Format and naming fix
