+++ PortunusAdiutor.csproj
@@ -23,7 +23,7 @@
     <RepositoryUrl>https://github.com/fabio-osti/TokenAuthenticationHelper.git</RepositoryUrl>
   </PropertyGroup>
 
-  <PropertyGroup>
+  <!-- <PropertyGroup>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
-  </PropertyGroup>
+  </PropertyGroup> -->
 </Project>
