+++ Services/MessagePoster/CodeMessagePoster.cs
@@ -1,12 +1,6 @@
-using System.Security.Claims;
-using System.Security.Cryptography;
+using Microsoft.AspNetCore.Identity;
 
 using MailKit.Net.Smtp;
-
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
-using Microsoft.IdentityModel.Tokens;
-
 using MimeKit;
 
 using PortunusAdiutor.Data;
@@ -16,9 +10,9 @@ using PortunusAdiutor.Services.TokenBuilder;
 namespace PortunusAdiutor.Services.MessagePoster;
 
 /// <summary>
-/// 	Implementation of <see cref="IMessagePoster{TUser, TKey}"/> with the plain text SUT.
+/// 	Implementation of <see cref="IMessagePoster{TUser, TKey}"/> with the plain text OTP.
 /// </summary>
-/// <typeparam name="TContext">Represents an Entity Framework database context used for identity with SUT keeping.</typeparam>
+/// <typeparam name="TContext">Represents an Entity Framework database context used for identity with OTP keeping.</typeparam>
 /// <typeparam name="TUser">Represents an user in the identity system.</typeparam>
 /// <typeparam name="TRole">Represents a role in the identity system.</typeparam>
 /// <typeparam name="TKey">Represents the key of an user in the identity system.</typeparam>
@@ -27,9 +21,9 @@ namespace PortunusAdiutor.Services.MessagePoster;
 /// <typeparam name="TUserLogin">Represents a login and its associated provider for an user.</typeparam>
 /// <typeparam name="TRoleClaim">Represents a claim that is granted to all users within a role.</typeparam>
 /// <typeparam name="TUserToken">Represents an authentication token for an user.</typeparam>
-public class MessageLinkPoster<TContext, TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken> : MessagePosterBase<TContext, TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken>, IMessagePoster<TUser, TKey>
+public class CodeMessagePoster<TContext, TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken> : MessagePosterBase<TContext, TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken>, IMessagePoster<TUser, TKey>
 where TContext : IdentityWithSutDbContext<TUser, TRole, TKey, TUserClaim, TUserRole, TUserLogin, TRoleClaim, TUserToken>
-where TUser : IdentityUser<TKey>, IManagedUser
+where TUser : IdentityUser<TKey>
 where TRole : IdentityRole<TKey>
 where TKey : IEquatable<TKey>
 where TUserClaim : IdentityUserClaim<TKey>
@@ -38,58 +32,52 @@ where TUserLogin : IdentityUserLogin<TKey>
 where TRoleClaim : IdentityRoleClaim<TKey>
 where TUserToken : IdentityUserToken<TKey>
 {
-	private readonly MessageLinkPosterParams _posterParams;
-	private readonly ITokenBuilder _tokenBuilder;
+	CodeMessagePosterParams _posterParams;
 
 	/// <summary>
 	/// 	Initializes an instance of the class.
 	/// </summary>
 	/// <param name="posterParams">Parameters for sending messages.</param>
 	/// <param name="context">Database context.</param>
-	/// <param name="tokenBuilder">JWT builder.</param>
-	public MessageLinkPoster(
-		MessageLinkPosterParams posterParams,
-		TContext context,
-		ITokenBuilder tokenBuilder
-	) : base(context)
+	public CodeMessagePoster(CodeMessagePosterParams posterParams, TContext context) : base(context)
 	{
 		_posterParams = posterParams;
-		_tokenBuilder = tokenBuilder;
 	}
 
 	/// <inheritdoc/>
 	public void SendEmailConfirmationMessage(TUser user)
 	{
-		// Generates SUT
-		var sut = GenAndSave(user.Id, MessageTypes.EmailConfirmation, out _);
-		// Builds and sends message
 		ArgumentException.ThrowIfNullOrEmpty(user.Email);
-		var message = _posterParams.PasswordRedefinitionMessageBuilder(
+
+		var otp = GenAndSave(user.Id, MessageTypes.EmailConfirmation, out var xdc);
+
+		var message = _posterParams.EmailConfirmationMessageBuilder(
 			user.Email,
-			_posterParams.PasswordRedefinitionEndpoint + sut.Token
+			xdc
 		);
+
 		SendMessage(message);
 	}
 
 	/// <inheritdoc/>
 	public void SendPasswordRedefinitionMessage(TUser user)
 	{
-		// Generates SUT
-		var sut = GenAndSave(user.Id, MessageTypes.PasswordRedefinition, out _);
-		// Builds and sends message
 		ArgumentException.ThrowIfNullOrEmpty(user.Email);
+
+		var otp = GenAndSave(user.Id, MessageTypes.PasswordRedefinition, out var xdc);
+
 		var message = _posterParams.PasswordRedefinitionMessageBuilder(
 			user.Email,
-			_posterParams.PasswordRedefinitionEndpoint + sut.Token
+			xdc
 		);
+
 		SendMessage(message);
 	}
 
 	/// <inheritdoc/>
 	private void SendMessage(MimeMessage message)
 	{
-		using (var client = new SmtpClient())
-		{
+		using (var client = new SmtpClient()) {
 			client.Connect(_posterParams.SmtpUri);
 			client.Authenticate(_posterParams.SmtpCredentials);
 			client.Send(message);
