+++ Services/MessagePoster/CodeMessagePosterParams.cs
@@ -11,7 +11,7 @@ namespace PortunusAdiutor.Services.MessagePoster;
 /// <summary>
 /// 	Parameters necessary for the link message posting.
 /// </summary>
-public class MessageLinkPosterParams
+public class CodeMessagePosterParams
 {
 	/// <summary>
 	/// 	Uri used for the SMTP server.
@@ -21,43 +21,30 @@ public class MessageLinkPosterParams
 	/// <summary>
 	/// 	Credentials used for the SMTP server.
 	/// </summary>
-	public ICredentials SmtpCredentials
-	{ get; set; } = new NetworkCredential();
-
-	/// <summary>
-	/// 	App endpoint for email validation.
-	/// </summary>
-	public string EmailConfirmationEndpoint { get; set; } =
-		defaultEmailConfirmationEndpoint;
-
-	/// <summary>
-	/// 	App endpoint for password redefinition.
-	/// </summary>
-	public string PasswordRedefinitionEndpoint { get; set; } =
-		defaultPasswordRedefinitionEndpoint;
-
+	public ICredentials SmtpCredentials{ get; set; } = defaultCredentials;
+		
 	/// <summary>
 	///		Sets or gets the builder of the email that should be sent if the user
 	///		forgets his password.
 	/// </summary>
-	public MessageBuilder PasswordRedefinitionMessageBuilder
-	{ get; set; } = defaultPasswordRedefinitionMessageBuilder;
+	public MessageBuilder PasswordRedefinitionMessageBuilder{ get; set; } = 
+		defaultPasswordRedefinitionMessageBuilder;
 
 	/// <summary>
 	///		Sets or gets the builder of the email that should be sent when the user 
 	///		is registered.
 	/// </summary>
-	public MessageBuilder EmailConfirmationMessageBuilder
-	{ get; set; } = defaultEmailConfirmationMessageBuilder;
+	public MessageBuilder EmailConfirmationMessageBuilder{ get; set; } = 
+		defaultEmailConfirmationMessageBuilder;
 
 	/// <summary>
-	/// 	Initialize an instance of <see cref="MessageLinkPosterParams"/>
+	/// 	Initialize an instance of <see cref="CodeMessagePosterParams"/>
 	/// 	with only the defaults as base.
 	/// </summary>
-	public MessageLinkPosterParams() {	}
+	public CodeMessagePosterParams() { }
 
 	/// <summary>
-	/// 	Iniatialize an instance of <see cref="MessageLinkPosterParams"/> 
+	/// 	Iniatialize an instance of <see cref="LinkMessagePosterParams"/> 
 	/// 	using an <see cref="IConfiguration"/> object and
 	/// 	the defaults as base.
 	/// </summary>
@@ -65,46 +52,28 @@ public class MessageLinkPosterParams
 	/// 	An <see cref="IConfiguration"/> instance that 
 	/// 	have the section "SMTP" defined.
 	/// </param>
-	public MessageLinkPosterParams(IConfiguration config)
+	public CodeMessagePosterParams(ConfigurationManager config)
 	{
 		var sect = config.GetSection("SMTP");
-		var smtpUri = sect["URI"];
+		var smtpUri = config["SMTP_URI"];
 		if (smtpUri is not null) {
 			SmtpUri = new(smtpUri);
 		}
 
-		var smtpUser = sect["USERNAME"];
+		var smtpUser = config["SMTP_USER"];
 		if (smtpUser is not null) {
-			var smtpPassword = sect["PSWRD"];
+			var smtpPassword = config["SMTP_PSWRD"];
 			SmtpCredentials =
 				new NetworkCredential(smtpUser, smtpPassword);
 		}
-
-		var emailConfirmationEndpoint = sect["ECE"];
-		if (emailConfirmationEndpoint is not null) {
-			EmailConfirmationEndpoint = emailConfirmationEndpoint;
-		}
-
-		var passwordRedefinitionEndpoint = sect["PRE"];
-		if (passwordRedefinitionEndpoint is not null) {
-			PasswordRedefinitionEndpoint = passwordRedefinitionEndpoint;
-		}
 	}
 
-	// DEFAULT VALUES
 	const string defaultSmtpUriString = "smtp://localhost:2525";
-
-	const string defaultEmailConfirmationEndpoint =
-		"http://localhost:8080/Authorization/ConfirmEmail?token=";
-
-	const string defaultPasswordRedefinitionEndpoint =
-		"http://localhost:8080/Authorization/RedefinePassword?token=";
-
 	static ICredentials defaultCredentials => new NetworkCredential();
 
 	static MimeMessage defaultPasswordRedefinitionMessageBuilder(
 		string email,
-		string link
+		string code
 	)
 	{
 		var message = new MimeMessage();
@@ -119,9 +88,9 @@ public class MessageLinkPosterParams
 
 				A new password was requested for your account,
 
-				Please confirm that it was you by opening this link: 
+				Please confirm that it was you by entering this code: 
 				
-				{link}
+				{code}
 
 				If you didn’t make this request, then you can ignore this email.
 				"""
@@ -132,7 +101,7 @@ public class MessageLinkPosterParams
 
 	static MimeMessage defaultEmailConfirmationMessageBuilder(
 		string email,
-		string link
+		string code
 	)
 	{
 		var message = new MimeMessage();
@@ -147,9 +116,9 @@ public class MessageLinkPosterParams
 
 				Your account have been registered, 
 
-				Please confirm that it was you by opening this link: 
+				Please confirm that it was you by entering this code: 
 
-				{link}
+				{code}
 
 				If you didn’t make this request, then you can ignore this email.
 				"""
