+++ Services/MessagePoster/LinkMessagePosterParams.cs
@@ -11,7 +11,7 @@ namespace PortunusAdiutor.Services.MessagePoster;
 /// <summary>
 /// 	Parameters necessary for the link message posting.
 /// </summary>
-public class MessageCodePosterParams
+public class LinkMessagePosterParams
 {
 	/// <summary>
 	/// 	Uri used for the SMTP server.
@@ -21,30 +21,42 @@ public class MessageCodePosterParams
 	/// <summary>
 	/// 	Credentials used for the SMTP server.
 	/// </summary>
-	public ICredentials SmtpCredentials
-	{ get; set; } = new NetworkCredential();
+	public ICredentials SmtpCredentials{ get; set; } = defaultCredentials;
+		
+	/// <summary>
+	/// 	App endpoint for email validation.
+	/// </summary>
+	public string EmailConfirmationEndpoint { get; set; } =
+		defaultEmailConfirmationEndpoint;
+
+	/// <summary>
+	/// 	App endpoint for password redefinition.
+	/// </summary>
+	public string PasswordRedefinitionEndpoint { get; set; } =
+		defaultPasswordRedefinitionEndpoint;
+
 	/// <summary>
 	///		Sets or gets the builder of the email that should be sent if the user
 	///		forgets his password.
 	/// </summary>
-	public MessageBuilder PasswordRedefinitionMessageBuilder
-	{ get; set; } = defaultPasswordRedefinitionMessageBuilder;
+	public MessageBuilder PasswordRedefinitionMessageBuilder{ get; set; } = 
+		defaultPasswordRedefinitionMessageBuilder;
 
 	/// <summary>
 	///		Sets or gets the builder of the email that should be sent when the user 
 	///		is registered.
 	/// </summary>
-	public MessageBuilder EmailConfirmationMessageBuilder
-	{ get; set; } = defaultEmailConfirmationMessageBuilder;
+	public MessageBuilder EmailConfirmationMessageBuilder{ get; set; } = 
+		defaultEmailConfirmationMessageBuilder;
 
 	/// <summary>
-	/// 	Initialize an instance of <see cref="MessageCodePosterParams"/>
+	/// 	Initialize an instance of <see cref="LinkMessagePosterParams"/>
 	/// 	with only the defaults as base.
 	/// </summary>
-	public MessageCodePosterParams() { }
+	public LinkMessagePosterParams() {	}
 
 	/// <summary>
-	/// 	Iniatialize an instance of <see cref="MessageLinkPosterParams"/> 
+	/// 	Iniatialize an instance of <see cref="LinkMessagePosterParams"/> 
 	/// 	using an <see cref="IConfiguration"/> object and
 	/// 	the defaults as base.
 	/// </summary>
@@ -52,28 +64,46 @@ public class MessageCodePosterParams
 	/// 	An <see cref="IConfiguration"/> instance that 
 	/// 	have the section "SMTP" defined.
 	/// </param>
-	public MessageCodePosterParams(ConfigurationManager config)
+	public LinkMessagePosterParams(IConfiguration config)
 	{
 		var sect = config.GetSection("SMTP");
-		var smtpUri = config["SMTP_URI"];
+		var smtpUri = sect["URI"];
 		if (smtpUri is not null) {
 			SmtpUri = new(smtpUri);
 		}
 
-		var smtpUser = config["SMTP_USER"];
+		var smtpUser = sect["USERNAME"];
 		if (smtpUser is not null) {
-			var smtpPassword = config["SMTP_PSWRD"];
+			var smtpPassword = sect["PSWRD"];
 			SmtpCredentials =
 				new NetworkCredential(smtpUser, smtpPassword);
 		}
+
+		var emailConfirmationEndpoint = sect["ECE"];
+		if (emailConfirmationEndpoint is not null) {
+			EmailConfirmationEndpoint = emailConfirmationEndpoint;
+		}
+
+		var passwordRedefinitionEndpoint = sect["PRE"];
+		if (passwordRedefinitionEndpoint is not null) {
+			PasswordRedefinitionEndpoint = passwordRedefinitionEndpoint;
+		}
 	}
 
+	// DEFAULT VALUES
 	const string defaultSmtpUriString = "smtp://localhost:2525";
+
+	const string defaultEmailConfirmationEndpoint =
+		"http://localhost:8080/Authorization/ConfirmEmail?token=";
+
+	const string defaultPasswordRedefinitionEndpoint =
+		"http://localhost:8080/Authorization/RedefinePassword?token=";
+
 	static ICredentials defaultCredentials => new NetworkCredential();
 
 	static MimeMessage defaultPasswordRedefinitionMessageBuilder(
 		string email,
-		string code
+		string link
 	)
 	{
 		var message = new MimeMessage();
@@ -88,9 +118,9 @@ public class MessageCodePosterParams
 
 				A new password was requested for your account,
 
-				Please confirm that it was you by entering this code: 
+				Please confirm that it was you by opening this link: 
 				
-				{code}
+				{link}
 
 				If you didn’t make this request, then you can ignore this email.
 				"""
@@ -101,7 +131,7 @@ public class MessageCodePosterParams
 
 	static MimeMessage defaultEmailConfirmationMessageBuilder(
 		string email,
-		string code
+		string link
 	)
 	{
 		var message = new MimeMessage();
@@ -116,9 +146,9 @@ public class MessageCodePosterParams
 
 				Your account have been registered, 
 
-				Please confirm that it was you by entering this code: 
+				Please confirm that it was you by opening this link: 
 
-				{code}
+				{link}
 
 				If you didn’t make this request, then you can ignore this email.
 				"""
